refactor(frontend): decompose Alerts.tsx into Alerts/ directory
